Further code analysis improvements
git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk@1517569 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-shade-plugin/src/main/java/org/apache/maven/plugins/shade/filter/MinijarFilter.java
@@ -206,7 +206,7 @@ public class MinijarFilter
     {
         int classesTotal = classesRemoved + classesKept;
         log.info(
-            "Minimized " + classesTotal + " -> " + classesKept + " (" + (int) ( 100 * classesKept / classesTotal )
+            "Minimized " + classesTotal + " -> " + classesKept + " (" + 100 * classesKept / classesTotal
                 + "%)" );
     }
 }
